Improved Wang's algorithm to run in more deterministic time
As the multivariate factorization algorithm (Wang's) was tested
for correctness for multitude of inputs using different system
configuration schemes (e.g. using gmpy as ground coefficients),
now we can lower bounds for evaluation points of variables,
when transforming multivariate problem to a univariate one.

Now the algorithm will try to find three evaluation sets, such
that all sets generate the same number of univariate factors
after substitution. If a new set generates smaller number of
factors, then all other are discarded and algorithm tries to
find to other configurations with the same number of factors.

If any of evaluation sets gives factor list of length one, then
the input multivariate polynomial is irreducible and no further
factorization is computed.

The first evaluation set tried is always [0,0,...,0], to test
if rare but very promising situation happens. Other evaluation
sets are generated at random, modulo some integer. At first the
algorithm generates sets modulo 1, i.e. in range [-1, 1], then
modulo 2, i.e. in range [-2,2] and so on. For each modulus only
5 evaluation sets are checked. If algorithm wasn't able to find
three configuration that minimize number of univariate factors
then it proceeds to the next modulus and continues searching.

When three sets were found, then the algorithm chooses the one
which generates, after substitution, the univariate polynomial
of smallest max-norm. This is important because EEZ lifting
algorithm is very sensitive to coefficient sizes. Univariate
factorizations are only a small part of the total procedure
execution time, so it might be convenient to try a few more
factorizations to minimize further coefficients.

If you want to experiment with parameters described above,
refer to sympy/polys/factortools.py and look for EEZ_*
constants.

If you get ExtraneousFactors (in fact NotImplementdError)
exception then it may have to causes:

1. EEZ_* parameters were set too low and algorithm was stuck
   on local minimum of the number of univariate factors. In
   this case try to run algorithm several times to verify
   the problem, because it may find a suitable evaluation.

2. Ground types don't work and their failure is masked with
   known algorithmic behaviour. In this case don't hope
   for a successful factorization.

Both cases show bugs which have to be fixed.