Improved polynomial unification functions
Unification is a process where given two polynomials with
(possibly) different generators and coefficient domains
we get equivalent polynomials with the same generators
and coefficient domain.

In simple cases, e.g. when coefficient domains don't overlap
with generators, unification rules are also simple. However,
consider the following situation:

In [1]: f = Poly(x, x, y)

In [2]: g = Poly(x, y, z)

In [3]: f
Out[3]: Poly(x, x, y, domain='ZZ')

In [4]: g
Out[4]: Poly(x, y, z, domain='ZZ[x]')

Lets assume we want to compute f+g. Now the question is what
should be the common generators and coefficient domain. One
option is to use ZZ[x] domain, but then we loose information
that x was a generator in f. The other option is to assume
(x, y, z) the generator set, but in this scenario we loose
information that x was a part of domain in g.

Current solution is to not allow such unifications at all:

In [5]: f+g
(...)
UnificationFailed: can't unify ZZ with ZZ[x], given (x, y, z) generators

This states that the common set of generators is (x, y, z) and
that it intersects with one of the initial coefficient domains.
Such behavior will make computing with polynomials less error
prone and easier to understand. If user wants to compute f+g
anyway, then he must do conversion of on of the input
polynomials manually.