Implemented function for computing minimal polynomials
The new function is minpoly() which uses Groebner basis techniques
to find the minimal polynomial of an algebraic number. The result
is a primitive polynomial in ZZ[x] of minimal degree, e.g.:

In [1]: a = sqrt(2) + sqrt(3) + sqrt(6)

In [2]: minpoly(a, x)
Out[2]:
                 2    4
-23 - 48⋅x - 22⋅x  + x

In [3]: a = 2*I + sqrt(2 + I)

In [4]: minpoly(a, x)
Out[4]:
              2    4
37 + 8⋅x + 4⋅x  + x

This function may fail to compute the minimal polynomial even if
its input is an algebraic number. This happens when there are
several candidates of a solution and SymPy fails to solve zero
equivalence problem when substituting input argument for `x`.