Allow computing Groebner bases without generators
We allow such behaviour in reduced(), so it's a natural
thing to allow the same behaviour in groebner(). Note
that order of generators does matter, so it's better
to be explicit when playing with Groebner bases.

Example:

In [1]: var('c,s')
Out[1]: (c, s)

In [2]: f = expand((1 - c**2)**5 * (1 - s**2)**5 * (c**2 + s**2)**10)

In [3]: groebner([f, c**2 + s**2 - 1])
Out[3]:
⎡      2    2     10      12       14       16      18    20⎤
⎣-1 + c  + s , - c   + 5⋅c   - 10⋅c   + 10⋅c   - 5⋅c   + c  ⎦