Added a function for fast polynomial construction
If one have an expression in non-expanded form and wants to
build a polynomial out of it, then the standard approach is
to use Poly class. Poly expands its argument by default and
constructs a polynomial. This can be slow, because expand()
is slow.

A new function, poly(), was implemented to overcome this
issue. It takes an expression as its only argument and
recursively tries to convert the given expression into
a polynomial without any expanding.

Example:

In [1]: f = 2*(x + (y + z)**2)**2

In [2]: expand(f)
Out[2]:
             2        2        2       2  2        3        3      4      4
8⋅x⋅y⋅z + 2⋅x  + 4⋅x⋅y  + 4⋅x⋅z  + 12⋅y ⋅z  + 8⋅y⋅z  + 8⋅z⋅y  + 2⋅y  + 2⋅z

In [3]: poly(f).as_basic()
Out[3]:
             2        2        2       2  2        3        3      4      4
8⋅x⋅y⋅z + 2⋅x  + 4⋅x⋅y  + 4⋅x⋅z  + 12⋅y ⋅z  + 8⋅y⋅z  + 8⋅z⋅y  + 2⋅y  + 2⋅z

Note that this is a preliminary implementation so not all
classes of expressions are currently handled by poly().