Use gmpy by default in polys and show types at init
Previously the default ground types were Python's int and Fraction
types (or SymPy's Rational on 2.4 / 2.5). The new default is gmpy,
if gmpy is available. If not SymPy will fallback to slower types
as previously.

To cut down on confusion what ground types are being used in a
particular session, the information about ground types that were
setup is now displayed in the welcome message in isympy. The text
is in a form: "types: something" (as it is done with cache).

If SymPy was able to use the same types for both ZZ and QQ domains,
then 'something' will be set to 'gmpy', 'python' or 'sympy'. If this
is not the case, e.g. in 2.4 w/o gmpy, then SymPy uses mixed ground
types and 'something' will be in a form: "ZZ_type/QQ_type". In the
example given it would be "types: python/sympy", i.e. ZZ.dtype is
int and QQ.dtype is Rational.

Examples:

1. Python 2.6 with gmpy

$ python2.6 bin/isympy -q
IPython console for SymPy 0.6.7-git (Python 2.6.2) (types: gmpy)

2. Python 2.6 w/o gmpy

$ SYMPY_GROUND_TYPES=python python2.6 bin/isympy -q
IPython console for SymPy 0.6.7-git (Python 2.6.2) (types: python)

3. Python 2.4 w/o gmpy

$ python2.4 bin/isympy -q
Python console for SymPy 0.6.7-git (Python 2.4.4) (types: python/sympy)