Implemented infallible complex root counting algorithm
The implemented algorithm is the algorithm due to Collins
and Krandick (1992) which is a purely rational (infallible)
approach to counting complex roots of a polynomial with
rational coefficients in a rectangle in the complex plane.

The algorithm does not use Sturm sequences as other complex
root counting algorithms and doesn't have problems with roots
located on the boundary of a rectangle.

Example:

Consider a polynomial

 f = (z - 1)*(z + 1)*(z**2 - 2)*(z - I)*(z + I))

in R = [-2 - I, 1 + I/2] rectangle. The polynomial has 6 six
roots all together and 4 of them are in R (with 2 located on
the boundary of R).

Lets count the roots in R using the new algorithm:

In [1]: from sympy.polys.rootisolation import *

In [2]: f = Poly((z - 1)*(z + 1)*(z**2 - 2)*(z - I)*(z + I))

In [3]: dup_count_complex_roots(f.rep.rep, (QQ(-2), QQ(-1)), (QQ(1), QQ(1,2)), ZZ)
Out[3]: 4