Added a symbol singleton equal to all commutative symbols
Often we can encounter functions, mostly concerning polynomial
manipulation, that require a single symbol to be passed in their
argument lists. It's convenient to not force user to pass that
symbol in certain situations, e.g. to save typing or to simplify
logic of coded that uses a function of this kind.

If a symbol is not given explicitly then a function needs to
construct a new one. For this purpose dummy symbols are used
to assure that the introduced symbol won't overlap with other
symbols in the context. This, however, leads to very unpleasant
behavior, e.g.:

In [1]: minpoly(sqrt(2)) == minpoly(sqrt(2))
Out[1]: False

To overcome this difficulty a symbol, which is equal to all
commutative symbols, was introduced, S.Pure or pure for short.

In [1]: pure == x
Out[1]: True

In [2]: x == pure
Out[2]: True

In [3]: pure**2 + sin(pure) - x*pure*I
Out[3]:
                   2
-ⅈ⋅x⋅℘ + sin(℘) + ℘

In [4]: _.atoms(Symbol)
Out[4]: set(x, ℘)

In future, functions like minpoly() will return an expression
with the pure symbol instead of dummy symbols. This will allow
more effective experimentation on the user level and simplification
of the interior of SymPy.